infrastructure/requirements.txt
aws-cdk.core==1.62.0
aws-cdk.aws_iam==1.62.0
aws-cdk.aws_sqs==1.62.0
aws-cdk.aws_sns==1.62.0
aws-cdk.aws_sns_subscriptions==1.62.0
aws-cdk.aws_s3==1.62.0
aws-cdk.aws_s3_notifications==1.62.0
aws-cdk.aws_lambda==1.62.0
aws-cdk.aws_lambda_event_sources==1.62.0
aws-cdk.aws_ecr==1.62.0
aws-cdk.aws_ecs==1.62.0
aws-cdk.aws_ec2==1.62.0
aws-cdk.aws_ssm==1.62.0
aws-cdk.aws_codepipeline==1.62.0
aws-cdk.aws_codepipeline_actions==1.62.0
aws-cdk.aws_cloudwatch==1.62.0
aws-cdk.aws_cloudwatch_actions==1.62.0
aws-cdk.aws_codebuild==1.62.0
aws-cdk.aws_events_targets==1.62.0
aws-cdk.aws_events==1.62.0
aws-cdk.aws_events_targets==1.62.0